Added Connected File - need to fill out formik
